Complete game tutorial

+++ pygame-tutorial/enemy.py
@@ -6,10 +6,10 @@ from pygame import Vector2
 class Enemy(GameObject):
     """ The Enemy class handles enemy movement, speed, etc """
     def __init__(self, pos: Vector2, scale: Vector2, image_path: str,
-                 speed: int) -> None:
+                 speed: float) -> None:
         """ Initializes the Player """
         super().__init__(pos, scale.x, scale.y, image_path)
-        self.speed: int = speed
+        self.speed: float = speed
 
     def move(self, max_width: float):
         """ Move the enemy """
